+++ app/Filters/LoginFilter.php
@@ -18,7 +18,7 @@ class LoginFilter implements FilterInterface
     public function before(RequestInterface $request, $arguments = null)
     {
         if (! session()->get('isLoggedIn')) {
-            return redirect()->to('home/login');
+            return redirect()->to('login');
         }
     }